src/requirements.txt
pypdf
langchain
sentence-transformers
faiss-cpu
pydantic
streamlit
